// src/models/data-md-penugasan.tsx
export interface ResponseDataMdPenugasan {
  data: DataMdPenugasan[]
  error: boolean
  pesan: any
}

export interface DataMdPenugasan {
  id_jabatan: any
  id_bidang: any
  id_bidangsub: any
  cap_offtask: any
  code_assign: any
  id_offpegawai: any
  id_offtask: any
  id_penerimatask: any
  jam_offtask: any
  name_offpegawai: any
  note_offtask: any
  reply_note: any
  reply_state: any
  str_file: any
  tgl_offtask: any
  user_assign: any
  plt_bidang: any
  plt_bidangsub: any
}